Fixes xruns on track loading on Raspberry Pi
When a track is not longer needed it now flags the memory as reusable
instead of calling the free() routine. Apparently the free() is
blocking the realtime thread causing xruns on computers low on CPU
resources like the Raspberry Pi (verified on version 3b), leading
to a short skip if a track is playing on the other deck.

TODO: Need to free() the memory on exit of the software.

TODO: Implement tests for loading multiple tracks, asserting a correct
behavior.